+++ server/manager/src/main/java/ge/ai/domino/server/manager/game/heuristic/RoundHeuristic.java
@@ -4,5 +4,5 @@ import ge.ai.domino.domain.game.Round;
 
 public interface RoundHeuristic {
 
-    float getHeuristic(Round round);
+    double getHeuristic(Round round);
 }
